// generated/src/app/api-models/cap-util-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapUtilWork {
  seqCcalcId: number;
  seqCapUtilWork: number;
  seqCapPoolId: number;
  applyTo: string;
  capStoplossId: string;
  seqClaimId: number;
  fileType: string;
  lineNumber: number;
  subLineCode: string;
  detailSvcDate: Date;
  seqCapElgcntWork: number;
  seqCapProvId: number;
  seqCapVendId: number;
  capEntityCode: string;
  seqProvId: number;
  seqMembId: number;
  procedureCode: string;
  billedAmt: number;
  allowedAmt: number;
  withholdAmt: number;
  netAmt: number;
  applicableAmt: number;
  claimStatus: string;
  processingStatus: string;
  medDefCode: string;
  postDate: Date;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  dateReceived: Date;
  companyCode: string;
  glRefCode: string;
  postingRule: string;
  seqCapVendAddress: number;
}